Time: 1 ms (94.68%), Space: 43.6 MB (16.59%) - LeetHub

+++ 0118-pascals-triangle/0118-pascals-triangle.java
@@ -7,12 +7,12 @@ class Solution {
         ans.add(firstrow);
         for(int i=2 ; i<=numRows ; i++){
             List<Integer> row = new ArrayList<>();
-            int a = 1;
-            row.add(a);
+            long a = 1;
+            row.add((int)a);
             for(int j=1 ; j<i ; j++){
                 a = a*(i-j);
                 a = a/(j);
-                row.add(a);
+                row.add((int)a);
             }
             ans.add(row);
         }
